Extract VirtualMachineInstance controller to own package
