+++ pkg/virt-controller/watch/vmi/vmi.go
@@ -16,7 +16,7 @@
  *
  */
 
-package watch
+package vmi
 
 import (
 	"context"
