+++ pkg/virt-controller/watch/vmi/vmi_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package watch
+package vmi
 
 import (
 	"context"
