+++ tests/BUILD.bazel
@@ -186,7 +186,7 @@ go_test(
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/leaderelectionconfig:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
-        "//pkg/virt-controller/watch:go_default_library",
+        "//pkg/virt-controller/watch/migration:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-handler:go_default_library",
         "//pkg/virt-handler/cgroup:go_default_library",
